+++ src/main/java/gs/mclo/api/internal/package-info.java
@@ -1,6 +1,6 @@
 @ApiStatus.Internal
 @NotNullByDefault
-package gs.mclo.api.util;
+package gs.mclo.api.internal;
 
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNullByDefault;