+++ src/main/java/gs/mclo/api/internal/request/UploadLogRequestBody.java
@@ -1,0 +1,20 @@
+package gs.mclo.api.internal.request;
+
+import gs.mclo.api.data.Metadata;
+import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.Set;
+
+@ApiStatus.Internal
+public final class UploadLogRequestBody {
+    private final String content;
+    private @Nullable String source;
+    private Set<Metadata<?>> metadata;
+
+    public UploadLogRequestBody(String content, @Nullable String source, Set<Metadata<?>> metadata) {
+        this.content = content;
+        this.source = source;
+        this.metadata = metadata;
+    }
+}