+++ src/main/java/gs/mclo/api/internal/request/package-info.java
@@ -1,0 +1,2 @@
+@org.jetbrains.annotations.NotNullByDefault
+package gs.mclo.api.internal.request;