+++ src/test/java/gs/mclo/api/ApiTest.java
@@ -1,0 +1,19 @@
+package gs.mclo.api;
+
+import org.junit.jupiter.api.BeforeEach;
+
+public class ApiTest {
+    protected MclogsClient client;
+
+    @BeforeEach
+    protected void setUp() {
+        client = new MclogsClient("aternos/mclogs-java-tests");
+
+        var apiUrl = System.getenv("MCLOGS_API_URL");
+        var logViewUrl = System.getenv("MCLOGS_LOG_VIEW_URL");
+
+        if ((apiUrl != null && !apiUrl.isBlank()) || (logViewUrl != null && !logViewUrl.isBlank())) {
+            client.setInstance(new Instance(apiUrl, logViewUrl));
+        }
+    }
+}