+++ src/test/java/gs/mclo/api/BaseApiTest.java
@@ -1,6 +1,0 @@
-package gs.mclo.api;
-
-public class BaseApiTest {
-    protected MclogsClient client = new MclogsClient("aternos/mclogs-java-tests");
-
-}